requirements.txt
streamlit
biopython
requests
pandas
plotly
kaleido
fpdf
python-dotenv
google-generativeai
google-adk
prometheus-client
pydantic
redis
sqlalchemy
tiktoken
colorama
psutil
matplotlib
